+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ☮️ Peace in the world, or the world in pieces. 🕊️
 ### Added
 - Added a hook (`'swimShapeChanged'`) that is called after the shape changer completed the shape changing process. It passes the new token in order to work with that further in code.
+- Dramatic Task Planner Macro. Read the Wiki on it, it has a couple features that are explained there.
 ### Fixed
 - Fixed chase deck table not resetting results when resetting a chase scene.